+++ app/src/main/java/mihon/core/migration/migrations/Migrations.kt
@@ -10,4 +10,5 @@ val migrations: List<Migration>
         CategoryPreferencesCleanupMigration(),
         InstallationIdMigration(),
         NovelRateLimitPreferencesMigration(),
+        VerticalNavigatorMigration(),
     )