+++ domain/src/main/java/tachiyomi/domain/manga/model/Manga.kt
@@ -1,14 +1,20 @@
 package tachiyomi.domain.manga.model
 
+import android.annotation.SuppressLint
 import androidx.compose.runtime.Immutable
 import eu.kanade.tachiyomi.source.model.SManga
 import eu.kanade.tachiyomi.source.model.UpdateStrategy
+import kotlinx.serialization.Serializable
+import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.JsonObject
 import mihon.core.common.extensions.EMPTY
 import tachiyomi.core.common.preference.TriState
-import java.io.Serializable
+import java.io.ObjectStreamException
 import java.time.Instant
+import java.io.Serializable as JavaSerializable
 
+@SuppressLint("UnsafeOptInUsageError")
+@Serializable
 @Immutable
 data class Manga(
     val id: Long,
@@ -38,7 +44,7 @@ data class Manga(
     val notes: String,
     val isNovel: Boolean = false,
     val memo: JsonObject,
-) : Serializable {
+) : JavaSerializable {
 
     val expectedNextUpdate: Instant?
         get() = nextUpdate
@@ -138,5 +144,18 @@ data class Manga(
             isNovel = false,
             memo = JsonObject.EMPTY,
         )
+    }
+
+    @Throws(ObjectStreamException::class)
+    private fun writeReplace(): Any {
+        return JavaToKotlinXSerializable(Json.encodeToString<Manga>(this))
+    }
+
+    class JavaToKotlinXSerializable(private val data: String) : JavaSerializable {
+
+        @Throws(ObjectStreamException::class)
+        private fun readResolve(): Any {
+            return Json.decodeFromString<Manga>(data)
+        }
     }
 }